fix: changed return type in insert operation and removed models from enities

+++ app/src/main/java/com/example/plantee/data/local/dao/DiagnosisDao.kt
@@ -14,7 +14,7 @@ import kotlinx.coroutines.flow.Flow
 @Dao
 interface DiagnosisDao {
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insert(diagnosis: DiagnosisEntity): Int
+    suspend fun insert(diagnosis: DiagnosisEntity): Long
 
     @Update
     suspend fun update(diagnosis: DiagnosisEntity)
